requirements.txt
streamlit>=1.28.0
firebase-admin>=6.0.0
google-genai>=0.3.0
spacy>=3.7.0,<3.8.0
groq>=0.4.0

chromadb>=0.4.0
sentence-transformers>=2.2.2
torch>=2.0.0

# Added for search functionality
ddgs>=4.2.0

# Added for background tasks
APScheduler>=3.10.0

# Added for environment variables
python-dotenv>=1.0.0

# Added for WordNet and spell-checking
nltk>=3.8.0
requests>=2.28.0